Joshua

+++ Caesar_3_MVC/Controller/MouseInputHandler.py
@@ -22,12 +22,11 @@ class MouseInputHandler:
         """
         Receive events posted to the message queue. 
         """
-
         if event.type == pygame.MOUSEBUTTONDOWN:
                 if event.button == 1: 
                         self.clicked = True
                         self.initialMouseCoordinate = pygame.mouse.get_pos()
-        if event.type == pygame.MOUSEBUTTONUP:
+        elif event.type == pygame.MOUSEBUTTONUP:
                 if(self.clicked):
                         # get current state
                         self.finalClickCoordinate = pygame.mouse.get_pos()
@@ -36,9 +35,14 @@ class MouseInputHandler:
                                 self.handleMouseEventsStateMenu(event)
                         if currentstate == STATE_PLAY:
                                 self.handleMouseEventsStatePlay(event)
-
                 if event.button == 1:
                         self.clicked = False
+                        self.initialMouseCoordinate = None
+        #  Preview clear land
+        elif event.type == pygame.MOUSEMOTION:
+            temp = pygame.mouse.get_pos()
+            if temp != self.initialMouseCoordinate and self.clicked:
+                self.handleMouseMouvement(event)
 
     def handleMouseEventsStateMenu(self, event):
         """
@@ -113,13 +117,18 @@ class MouseInputHandler:
 
             for xi in range(grid_x1, grid_x2+1):
                 for yi in range(grid_y1, grid_y2+1):
-                        self.model.actualGame.map[xi][yi].sprite = "land1"
-                        self.model.actualGame.map[xi][yi].collision = 0
-                        if self.model.actualGame.map[xi][yi].road :
-                            self.model.actualGame.map[xi][yi].road.delete()
-                        if self.model.actualGame.map[xi][yi].building :
-                            self.model.actualGame.map[xi][yi].building.delete()
+                        if self.model.actualGame.map[xi][yi].sprite not in list_of_undestructible:
+                          self.model.actualGame.map[xi][yi].sprite = "land1"
+                          self.model.actualGame.map[xi][yi].collision = 0
+                          if self.model.actualGame.map[xi][yi].road :
+                              self.model.actualGame.map[xi][yi].road.delete()
+                          if self.model.actualGame.map[xi][yi].building :
+                              self.model.actualGame.map[xi][yi].building.delete()
+
             self.model.actualGame.collision_matrix = self.model.actualGame.create_collision_matrix()
+            for xi in range(len(self.model.actualGame.previewMap)):
+                for yi in range(len(self.model.actualGame.previewMap[0])):
+                        self.model.actualGame.previewMap[xi][yi] = None
         
                 
         #Routes
@@ -199,8 +208,8 @@ class MouseInputHandler:
                             Route(self.model.actualGame.map[grid_x1][yi], self.model.actualGame)
                             
             self.model.actualGame.collision_matrix = self.model.actualGame.create_collision_matrix()
-
-        #Building
+            
+             #Building
         if build_housing_button.clicked and not build_housing_button.rect.collidepoint(event.pos):
             x, y = self.initialMouseCoordinate
             world_x = x - self.model.actualGame.camera.vect.x - self.model.actualGame.surface_cells.get_width() / 2
@@ -250,5 +259,65 @@ class MouseInputHandler:
 
             for xi in range(grid_x1, grid_x2+1):
                 for yi in range(grid_y1, grid_y2+1):
-                    if not self.model.actualGame.map[xi][yi].road and not self.model.actualGame.map[xi][yi].building:
-                        Building((xi, yi), self.model.actualGame.map[xi][yi], self.model.actualGame, 1, False, 0, 0)
+                   if not self.model.actualGame.map[xi][yi].road and not self.model.actualGame.map[xi][yi].building:
+                          Building((xi, yi), self.model.actualGame.map[xi][yi], self.model.actualGame, 1, False, 0, 0)
+
+    def handleMouseMouvement(self, event):
+        """ Here we are going to manage the movement of the mouse"""
+        #Pelle
+        if clear_land_button.clicked and not clear_land_button.rect.collidepoint(event.pos) and self.initialMouseCoordinate != None:
+            for xi in range(len(self.model.actualGame.previewMap)):
+                for yi in range(len(self.model.actualGame.previewMap[0])):
+                        self.model.actualGame.previewMap[xi][yi] = None
+                        x, y = self.initialMouseCoordinate
+            world_x = x - self.model.actualGame.camera.vect.x - self.model.actualGame.surface_cells.get_width() / 2
+            world_y = y - self.model.actualGame.camera.vect.y
+
+            cart_y = (2 * world_y - world_x) / 2
+            cart_x = cart_y + world_x
+            grid_x1 = int(cart_x // cell_size)
+            grid_y1 = int(cart_y // cell_size)
+
+            x, y = event.pos
+            world_x = x - self.model.actualGame.camera.vect.x - self.model.actualGame.surface_cells.get_width() / 2
+            world_y = y - self.model.actualGame.camera.vect.y
+
+            cart_y = (2 * world_y - world_x) / 2
+            cart_x = cart_y + world_x
+            grid_x2 = int(cart_x // cell_size)
+            grid_y2 = int(cart_y // cell_size)
+        
+            if grid_x1 <0:
+                grid_x1 = 0
+            if grid_x2 <0:
+                grid_x2 = 0
+            if grid_y1 <0:
+                grid_y1 = 0
+            if grid_y2 <0:
+                grid_y2 = 0
+
+            if grid_x1 > self.model.actualGame.nbr_cell_x-1:
+                grid_x1 = self.model.actualGame.nbr_cell_x-1
+            if grid_x2 > self.model.actualGame.nbr_cell_x-1:
+                grid_x2 = self.model.actualGame.nbr_cell_x-1
+            if grid_y1 > self.model.actualGame.nbr_cell_y-1:
+                grid_y1 = self.model.actualGame.nbr_cell_y-1
+            if grid_y2 > self.model.actualGame.nbr_cell_y-1:
+                grid_y2 = self.model.actualGame.nbr_cell_y-1
+
+            if grid_x1 > grid_x2:
+                temp = grid_x1
+                grid_x1 = grid_x2
+                grid_x2 = temp
+
+            if grid_y1 > grid_y2:
+                temp = grid_y1
+                grid_y1 = grid_y2
+                grid_y2 = temp
+
+            for xi in range(grid_x1, grid_x2+1):
+                for yi in range(grid_y1, grid_y2+1):
+                    self.model.actualGame.previewMap[xi][yi] = "red"
+
+=======
+
